Extract shared incremental-restore test helpers into ut_helpers/
Move WaitForIncrementalBackupDone and related polling/observer helpers
out of ut_incr_restore_reboots and into a new test_incremental_restore.h
under ut_helpers, shared with ut_incremental_restore. Lets the backup
tracking branch reuse the same primitives instead of forking them.
